requirements.txt
discord.py==1.7.3
opensea-api==0.1.3